Use textparse in prometheus module helper library
